:art: modal => factorize code
